Fix #234 - Pickup vscode's language specific editor settings.
Vscode's language specific settings are ignored because the extension
didn't pass the configuration scope while it invokes
`workspace.getConfiguration(...)` vscode API to retrieve the default
editor settings.

This commit fixes the issue by passing the target `TextDocument`
instance as the configuration scope to the vscode API to get the
correct editor settings.